flight/requirements.txt
pymongo[srv]==3.11
selenium==4.19.0
webdriver-manager==4.0.1
apache-airflow-providers-mongo==3.0.0